requirements.txt
tensorflow==2.8.0
numpy==1.21.2
matplotlib==3.4.3
Pillow==8.4.0
scikit-image==0.18.3
jupyter==1.0.0
opencv-python==4.5.3.20210927